many refactorings and optimisations / moving to MHV_AVR_Tools_20121007 (huge speedup from LTO!)

+++ common/display.h
@@ -8,7 +8,7 @@ void sevenSeg_setNumber(int8_t n);
 
 int led_getOn(p600LED_t led);
 int led_getBlinking(p600LED_t led);
-void led_set(p600LED_t led, int on, int blinking);
+void led_set(p600LED_t led, int8_t on, int8_t blinking);
 
 void display_init(void);
 void display_update(void);
